requirements.txt
opencv-python>=4.7
pillow
matplotlib
python-socketio
eventlet
websockets
